refactor(notifications): extract telegram channel into its own module
Per review on #3981: pull the telegram summary-section budgeting and
delivery branch out of deliver_background_notifications into a
dedicated telegram_channel.py, mirroring separation of concerns
already used elsewhere. notifications.py now just delegates to
deliver_telegram_notification, keeping the lazy-import boundary intact.
